+++ src/places/place_search/nearby_search/request/build.rs
@@ -1,11 +1,11 @@
-use crate::places::place_search::text_search::request::Request;
+use crate::places::place_search::nearby_search::request::Request;
 use percent_encoding::{utf8_percent_encode, NON_ALPHANUMERIC};
 
 // -----------------------------------------------------------------------------
 
 impl<'a> Request<'a> {
 
-    /// Builds the query string for the Google Maps Places API _Text Search_
+    /// Builds the query string for the Google Maps Places API _Nearby Search_
     /// query based on the input provided by the client.
     ///
     /// ## Arguments:
@@ -18,24 +18,23 @@ impl<'a> Request<'a> {
         // string:
 
         let mut query = format!(
-            "key={}&query={}&radius={}",
+            "key={}&location={}&radius={}",
             self.client.key,
-            utf8_percent_encode(&self.input, NON_ALPHANUMERIC),
+            String::from(&self.location),
             self.radius,
         );
+
         // This section builds the "optional parameters" portion of the query
         // string:
 
-        // Language key/value pair:
+        if let Some(keyword) = &self.keyword {
+            query.push_str("&keyword={}");
+            query.push_str(&utf8_percent_encode(&keyword, NON_ALPHANUMERIC).to_string())
+        }
+
         if let Some(language) = &self.language {
             query.push_str("&language=");
             query.push_str(&String::from(language))
-        }
-
-        // Location key/value pair:
-        if let Some(location) = &self.location {
-            query.push_str("&location=");
-            query.push_str(&String::from(location))
         }
 
         if let Some(maxprice) = &self.maxprice {
@@ -52,19 +51,16 @@ impl<'a> Request<'a> {
             if *opennow { query.push_str("&opennow") }
         }
 
-        // Page Token key/value pair:
         if let Some(pagetoken) = &self.pagetoken {
             query.push_str("&pagetoken=");
             query.push_str(pagetoken)
         }
 
-        // Region key/value pair:
-        if let Some(region) = &self.region {
-            query.push_str("&region=");
-            query.push_str(&String::from(region))
+        if let Some(rankby) = &self.rankby {
+            query.push_str("&rankby=");
+            query.push_str(&String::from(rankby))
         }
 
-        // Place Type key/value pair:
         if let Some(place_type) = &self.place_type {
             query.push_str("&type=");
             query.push_str(&String::from(place_type))