+++ src/places/place_search/nearby_search/request/execute.rs
@@ -1,0 +1,33 @@
+use crate::error::Error as GoogleMapsError;
+use crate::places::place_search::nearby_search::{
+    request::Request as TextSearchRequest,
+    response::Response as TextSearchResponse,
+}; // crate::places::place_search::nearby_search
+
+// =============================================================================
+
+impl<'a> TextSearchRequest<'a> {
+
+    // -------------------------------------------------------------------------
+    //
+    /// Executes the query you've built.
+    ///
+    /// ## Description:
+    ///
+    /// My adventures in Rust became messy so I had to make this method. It
+    /// wraps the `.validate()?.build()?.get()?` chain needed at the end of the
+    /// builder pattern.
+    ///
+    /// ## Arguments:
+    ///
+    /// This method accepts no arguments.
+
+    pub async fn execute(
+        &'a mut self
+    ) -> Result<TextSearchResponse, GoogleMapsError> {
+
+        self.build().get().await
+
+    } // fn
+
+} // impl