+++ src/places/place_search/nearby_search/request/mod.rs
@@ -1,5 +1,5 @@
 //! **Look in this module for documentation on building your _Places API_
-//! _Text Search_ query**. In particular, look at the _Request_ struct
+//! _Nearby Search_ query**. In particular, look at the _Request_ struct
 //! for examples of the builder pattern. This module contains the tools (enums,
 //! structs, methods) for building your Google Maps Platform request.
 
@@ -11,22 +11,24 @@ mod get;
 mod is_open_now;
 mod new;
 mod query_url;
+mod with_keyword;
 mod with_language;
-mod with_location;
 mod with_max_price;
 mod with_min_price;
 mod with_pagetoken;
-mod with_region;
+mod with_rankby;
 mod with_type;
 
 // -----------------------------------------------------------------------------
 
-use crate::{client::GoogleMapsClient, types::Language, types::LatLng, types::PlaceType, types::Region};
+use crate::LatLng;
+use crate::places::RankBy;
+use crate::{client::GoogleMapsClient, types::Language, types::PlaceType};
 
 // -----------------------------------------------------------------------------
 
 /// **Look at this `Request` struct for documentation on how to build your
-/// _Text Search_ query**. The methods implemented for this struct are
+/// _Nearby Search_ query**. The methods implemented for this struct are
 /// what's used to build your request.
 
 #[derive(Debug)]
@@ -38,17 +40,13 @@ pub struct Request<'a> {
     /// user-definable settings such as "maximum retries."
     client: &'a GoogleMapsClient,
 
-    /// The text string on which to search, for example: "restaurant" or "123
-    /// Main Street". This must a place name, address, or category of
-    /// establishments. Any other types of input can generate errors and are not
-    /// guaranteed to return valid results. The Google Places service will
-    /// return candidate matches based on this string and order the results
-    /// based on their perceived relevance.
-    input: String,
+    /// The point around which to retrieve place information. This must be
+    /// specified as `latitude,longitude`.
+    location: LatLng,
 
     /// Defines the distance (in meters) within which to return place results.
-    /// You may bias results to a specified circle by passing a location and a
-    /// radius parameter. Doing so instructs the Places service to prefer
+    /// You may bias results to a specified circle by passing a `location` and a
+    /// `radius` parameter. Doing so instructs the Places service to prefer
     /// showing results within that circle; results outside of the defined area
     /// may still be displayed.
     ///
@@ -64,11 +62,26 @@ pub struct Request<'a> {
     ///         * When using `rankby=distance`, the radius parameter will not be
     ///         accepted, and will result in an `INVALID_REQUEST`.
     /// * Query Autocomplete: 50,000 meters
-    /// * Text Search: 50,000 meters
+    /// * Nearby Search: 50,000 meters
     radius: u32,
 
     // Optional parameters:
     // --------------------
+
+    /// The text string on which to search, for example: "restaurant" or "123
+    /// Main Street". This must be a place name, address, or category of
+    /// establishments. Any other types of input can generate errors and are not
+    /// guaranteed to return valid results. The Google Places service will
+    /// return candidate matches based on this string and order the results
+    /// based on their perceived relevance.
+    ///
+    /// Explicitly including location information using this parameter may
+    /// conflict with the location, radius, and rankby parameters, causing
+    /// unexpected results.
+    ///
+    /// If this parameter is omitted, places with a business_status of
+    /// `CLOSED_TEMPORARILY` or `CLOSED_PERMANENTLY` will not be returned.
+    keyword: Option<String>,
 
     /// The language in which to return results.
     ///
@@ -97,13 +110,6 @@ pub struct Request<'a> {
     /// _tér_ are synonyms for street in Hungarian.
     language: Option<Language>,
 
-    /// The point around which to retrieve place information.
-    ///
-    /// * When using the Text Search API, the 'location' parameter may be
-    /// overriden if the 'query' contains an explicit location such as
-    /// 'Market in Barcelona'.
-    location: Option<LatLng>,
-
     /// Restricts results to only those places within the specified range. Valid
     /// values range between 0 (most affordable) to 4 (most expensive),
     /// inclusive. The exact amount indicated by a specific value will vary from
@@ -127,13 +133,20 @@ pub struct Request<'a> {
     /// previously — all parameters other than pagetoken will be ignored.
     pagetoken: Option<String>,
 
-    /// The region code, specified as a [ccTLD ("top-level
-    /// domain")](https://en.wikipedia.org/wiki/List_of_Internet_top-level_domains#Country_code_top-level_domains)
-    /// two-character value. Most ccTLD codes are identical to ISO 3166-1 codes,
-    /// with some notable exceptions. For example, the United Kingdom's ccTLD is
-    /// "uk" (.co.uk) while its ISO 3166-1 code is "gb" (technically for the
-    /// entity of "The United Kingdom of Great Britain and Northern Ireland").
-    region: Option<Region>,
+    /// Specifies the order in which results are listed. Possible values are:
+    ///
+    /// * `prominence` (default). This option sorts results based on their
+    /// importance. Ranking will favor prominent places within the set radius
+    /// over nearby places that match but that are less prominent. Prominence
+    /// can be affected by a place's ranking in Google's index, global
+    /// popularity, and other factors. When prominence is specified, the
+    /// `radius` parameter is required.
+    ///
+    /// * `distance`. This option biases search results in ascending order by
+    /// their distance from the specified location. When `distance` is
+    /// specified, one or more of `keyword`, `name`, or `type` is required and
+    /// radius is disallowed.
+    rankby: Option<RankBy>,
 
     /// Restricts the results to places matching the specified type. Only one
     /// type may be specified. If more than one type is provided, all types