+++ src/places/place_search/nearby_search/request/new.rs
@@ -1,5 +1,6 @@
 use crate::client::GoogleMapsClient;
-use crate::places::place_search::text_search::request::Request;
+use crate::LatLng;
+use crate::places::place_search::nearby_search::request::Request;
 
 // =============================================================================
 
@@ -7,19 +8,15 @@ impl<'a> Request<'a> {
 
     // -------------------------------------------------------------------------
     //
-    /// Initializes the builder pattern for a Places API _Text Search_ query
+    /// Initializes the builder pattern for a Places API _Nearby Search_ query
     /// with the required, non-optional parameters.
     ///
     /// ## Arguments:
     ///
     /// * `client` ‧ Your application's Google Maps API client struct.
     ///
-    /// * `query` ‧ The text string on which to search, for example:
-    /// "restaurant" or "123 Main Street". This must a place name, address, or
-    /// category of establishments. Any other types of input can generate errors
-    /// and are not guaranteed to return valid results. The Google Places
-    /// service will return candidate matches based on this string and order the
-    /// results based on their perceived relevance.
+    /// * `location` ‧ The point around which to retrieve place information.
+    /// This must be specified as `latitude,longitude`.
     ///
     /// * `radius` ‧ Defines the distance (in meters) within which to return
     /// place results. You may bias results to a specified circle by passing a
@@ -43,7 +40,7 @@ impl<'a> Request<'a> {
 
     pub fn new(
         client: &GoogleMapsClient,
-        query: String,
+        location: LatLng,
         radius: u32,
     ) -> Request {
 
@@ -51,16 +48,16 @@ impl<'a> Request<'a> {
         Request {
             // Required parameters:
             client,
-            input: query,
+            location,
             radius,
             // Optional parameters:
+            keyword: None,
             language: None,
-            location: None,
             maxprice: None,
             minprice: None,
             opennow: None,
             pagetoken: None,
-            region: None,
+            rankby: None,
             place_type: None,
             // Internal use only:
             query: None,